revise stack array test
