+++ spellbook/ml/learn.py
@@ -31,30 +31,12 @@
 import numpy as np
 
 import spellbook.ml.surrogates as surrogates
-
+from spellbook.utils import load_infile
 
 try:
     import cPickle as pickle
 except ImportError:
     import pickle
-
-
-def load_infile(args):
-    with np.load(args.infile) as data:
-        if args.X is not None:
-            X = stack_arrays(data, args.X)  # inputs
-        elif "X" in data.keys():
-            X = data["X"]
-        else:
-            X = data[data.files[0]]
-
-        if args.y is not None:
-            y = stack_arrays(data, args.y)  # outputs
-        elif "y" in data.keys():
-            y = data["y"]
-        else:
-            y = data[data.files[1]]
-    return X, y
 
 
 def make_regressor(args):
@@ -81,8 +63,3 @@ def make_regressor(args):
     regr.fit(X, y)
     with open(args.outfile, "wb") as f:
         pickle.dump(regr, f)
-
-
-def stack_arrays(data, delimited_names, delimiter=","):
-    stacked = np.vstack([data[name] for name in delimited_names.split(delimiter)])
-    return stacked.T
