+++ spellbook/ml/learn_alt.py
@@ -1,6 +1,6 @@
 import numpy as np
 from sklearn.ensemble import RandomForestRegressor
-
+from spellbook.utils import load_infile
 
 try:
     import cPickle as pickle
@@ -9,24 +9,6 @@ except ImportError:
 
 
 FOREST_DEFAULTS = {"max_depth": 2, "random_state": 0, "n_estimators": 100}
-
-
-def load_infile(args):
-    with np.load(args.infile) as data:
-        if args.X is not None:
-            X = stack_arrays(data, args.X)  # inputs
-        elif "X" in data.keys():
-            X = data["X"]
-        else:
-            X = data[data.files[0]]
-
-        if args.y is not None:
-            y = stack_arrays(data, args.y)  # outputs
-        elif "y" in data.keys():
-            y = data["y"]
-        else:
-            y = data[data.files[1]]
-    return X, y
 
 
 def random_forest(args):
@@ -50,8 +32,3 @@ def random_forest(args):
     regr.fit(X, y)
     with open(args.outfile, "wb") as f:
         pickle.dump(regr, f)
-
-
-def stack_arrays(data, delimited_names, delimiter=","):
-    stacked = np.vstack([data[name] for name in delimited_names.split(delimiter)])
-    return stacked.T
